fixed budgetform redirect, change graphql values from int => float, d…

+++ client/src/utils/mutations.js
@@ -49,7 +49,7 @@ export const ADD_EXPENSE = gql`
   mutation Mutation(
     $budgetId: ID!
     $name: String!
-    $cost: Int!
+    $cost: Float!
     $categoryId: ID!
   ) {
     addExpense(
@@ -76,7 +76,7 @@ export const ADD_EXPENSE = gql`
 `;
 
 export const ADD_INCOME = gql`
-  mutation AddIncome($budgetId: ID!, $name: String!, $amount: Int!) {
+  mutation AddIncome($budgetId: ID!, $name: String!, $amount: Float!) {
     addIncome(budgetId: $budgetId, name: $name, amount: $amount) {
       _id
       budgetMonth
@@ -138,7 +138,7 @@ export const UPDATE_EXPENSE = gql`
     $budgetId: ID!
     $expenseId: ID!
     $name: String
-    $cost: Int
+    $cost: Float
     $categoryId: ID
   ) {
     updateExpense(
@@ -170,7 +170,7 @@ export const UPDATE_INCOME = gql`
     $budgetId: ID!
     $incomeId: ID!
     $name: String
-    $amount: Int
+    $amount: Float
   ) {
     updateIncome(
       budgetId: $budgetId
